+++ .eslintrc.json
@@ -5,7 +5,7 @@
     "es6": true
   },
   "parserOptions": {
-    "ecmaVersion": 9
+    "ecmaVersion": 2019
   },
   "plugins": [
     "jsdom-internal"
