+++ lib/jsdom/living/nodes/HTMLInputElement-impl.js
@@ -22,22 +22,13 @@ const {
   asciiCaseInsensitiveMatch,
   splitOnCommas
 } = require("../helpers/strings");
+const { isDate } = require("../helpers/dates-and-times");
 const {
-  isDate,
-
-  parseDateString,
-  parseLocalDateAndTimeString,
-  parseMonthString,
-  parseTimeString,
-  parseWeekString,
-
-  serializeDate,
-  serializeMonth,
-  serializeNormalizedDateAndTime,
-  serializeTime,
-  serializeWeek,
-  parseDateAsWeek
-} = require("../helpers/dates-and-times");
+  convertStringToNumberByType,
+  convertStringToDateByType,
+  serializeDateByType,
+  convertNumberToStringByType
+} = require("../helpers/number-and-date-inputs");
 
 const filesSymbol = Symbol("files");
 
@@ -100,209 +91,6 @@ function valueAttributeMode(type) {
   return "value";
 }
 
-// Necessary because Date.UTC() treats year within [0, 99] as [1900, 1999].
-function getUTCMs(year, month = 1, day = 1, hour = 0, minute = 0, second = 0, millisecond = 0) {
-  if (year > 99 || year < 0) {
-    return Date.UTC(year, month - 1, day, hour, minute, second, millisecond);
-  }
-  const d = new Date(0);
-  d.setUTCFullYear(year);
-  d.setUTCMonth(month - 1);
-  d.setUTCDate(day);
-  d.setUTCHours(hour);
-  d.setUTCMinutes(minute);
-  d.setUTCSeconds(second, millisecond);
-  return d.valueOf();
-}
-
-const dayOfWeekRelMondayLUT = [-1, 0, 1, 2, 3, -3, -2];
-
-const convertStringToNumberByTypeMap = new Map([
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#date-state-(type=date):concept-input-value-string-number
-    "date", input => {
-      const date = parseDateString(input);
-      if (date === null) {
-        return NaN;
-      }
-      return getUTCMs(date.year, date.month, date.day);
-    }
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#month-state-(type=month):concept-input-value-string-number
-    "month", input => {
-      const date = parseMonthString(input);
-      if (date === null) {
-        return NaN;
-      }
-      return (date.year - 1970) * 12 + (date.month - 1);
-    }
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#week-state-(type=week):concept-input-value-string-number
-    "week", input => {
-      const date = parseWeekString(input);
-      if (date === null) {
-        return NaN;
-      }
-      const dateObj = new Date(getUTCMs(date.year));
-      // An HTML week starts on Monday, while 0 represents Sunday. Account for such.
-      const dayOfWeekRelMonday = dayOfWeekRelMondayLUT[dateObj.getUTCDay()];
-      return dateObj.setUTCDate(1 + 7 * (date.week - 1) - dayOfWeekRelMonday);
-    }
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#time-state-(type=time):concept-input-value-string-number
-    "time", input => {
-      const time = parseTimeString(input);
-      if (time === null) {
-        return NaN;
-      }
-      return ((time.hour * 60 + time.minute) * 60 + time.second) * 1000 + time.millisecond;
-    }
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#local-date-and-time-state-(type=datetime-local):concept-input-value-string-number
-    "datetime-local", input => {
-      const dateAndTime = parseLocalDateAndTimeString(input);
-      if (dateAndTime === null) {
-        return NaN;
-      }
-      const { date: { year, month, day }, time: { hour, minute, second, millisecond } } = dateAndTime;
-      // Doesn't quite matter whether or not UTC is used, since the offset from 1970-01-01 local time is returned.
-      return getUTCMs(year, month, day, hour, minute, second, millisecond);
-    }
-  ],
-  // https://html.spec.whatwg.org/multipage/input.html#number-state-(type=number):concept-input-value-string-number
-  ["number", parseFloatingPointNumber],
-  // https://html.spec.whatwg.org/multipage/input.html#range-state-(type=range):concept-input-value-string-number
-  ["range", parseFloatingPointNumber]
-]);
-
-const convertStringToDateByTypeMap = new Map([
-  [
-    "date", input => {
-      const parsedInput = convertStringToNumberByTypeMap.get("date")(input);
-      return isNaN(parsedInput) ? NaN : new Date(parsedInput);
-    }
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#month-state-(type=month):concept-input-value-string-number
-    "month", input => {
-      const parsedMonthString = parseMonthString(input);
-      if (parsedMonthString === null) {
-        return NaN;
-      }
-
-      const date = new Date(0);
-      date.setUTCFullYear(parsedMonthString.year);
-      date.setUTCMonth(parsedMonthString.month - 1);
-      return date;
-    }
-  ],
-  [
-    "week", input => {
-      const parsedInput = convertStringToNumberByTypeMap.get("week")(input);
-      return isNaN(parsedInput) ? NaN : new Date(parsedInput);
-    }
-  ],
-  [
-    "time", input => {
-      const parsedInput = convertStringToNumberByTypeMap.get("time")(input);
-      return isNaN(parsedInput) ? NaN : new Date(parsedInput);
-    }
-  ],
-  [
-    "datetime-local", input => {
-      const parsedInput = convertStringToNumberByTypeMap.get("datetime-local")(input);
-      return isNaN(parsedInput) ? NaN : new Date(parsedInput);
-    }
-  ]
-]);
-
-const serializeDateByTypeMap = new Map([
-  [
-    "date", input => {
-      return serializeDate({
-        year: input.getUTCFullYear(),
-        month: input.getUTCMonth() + 1,
-        day: input.getUTCDate()
-      });
-    }
-  ],
-  [
-    "month", input => {
-      return serializeMonth({
-        year: input.getUTCFullYear(),
-        month: input.getUTCMonth() + 1
-      });
-    }
-  ],
-  ["week", input => serializeWeek(parseDateAsWeek(input))],
-  [
-    "time", input => {
-      return serializeTime({
-        hour: input.getUTCHours(),
-        minute: input.getUTCMinutes(),
-        second: input.getUTCSeconds(),
-        millisecond: input.getUTCMilliseconds()
-      });
-    }
-  ],
-  [
-    "datetime-local", input => {
-      return serializeNormalizedDateAndTime({
-        date: {
-          year: input.getUTCFullYear(),
-          month: input.getUTCMonth() + 1,
-          day: input.getUTCDate()
-        },
-        time: {
-          hour: input.getUTCHours(),
-          minute: input.getUTCMinutes(),
-          second: input.getUTCSeconds(),
-          millisecond: input.getUTCMilliseconds()
-        }
-      });
-    }
-  ]
-]);
-
-const convertNumberToStringByTypeMap = new Map([
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#date-state-(type=date):concept-input-value-string-number
-    "date", input => serializeDateByTypeMap.get("date")(new Date(input))
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#month-state-(type=month):concept-input-value-string-date
-    "month", input => {
-      const year = 1970 + Math.floor(input / 12);
-      const month = input % 12;
-      const date = new Date();
-      date.setUTCFullYear(year);
-      date.setUTCMonth(month);
-
-      return serializeDateByTypeMap.get("month")(date);
-    }
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#week-state-(type=week):concept-input-value-string-date
-    "week", input => serializeDateByTypeMap.get("week")(new Date(input))
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#time-state-(type=time):concept-input-value-string-date
-    "time", input => serializeDateByTypeMap.get("time")(new Date(input))
-  ],
-  [
-    // https://html.spec.whatwg.org/multipage/input.html#local-date-and-time-state-(type=datetime-local):concept-input-value-number-string
-    "datetime-local", input => serializeDateByTypeMap.get("datetime-local")(new Date(input))
-  ],
-  // https://html.spec.whatwg.org/multipage/input.html#number-state-(type=number):concept-input-value-number-string
-  ["number", v => v.toString()],
-  // https://html.spec.whatwg.org/multipage/input.html#range-state-(type=range):concept-input-value-number-string
-  ["range", v => v.toString()]
-]);
-
 class HTMLInputElementImpl extends HTMLElementImpl {
   constructor(globalObject, args, privateData) {
     super(globalObject, args, privateData);
@@ -327,19 +115,19 @@ class HTMLInputElementImpl extends HTMLElementImpl {
 
   // https://html.spec.whatwg.org/multipage/input.html#concept-input-value-string-number
   get _convertStringToNumber() {
-    return convertStringToNumberByTypeMap.get(this.type);
+    return convertStringToNumberByType[this.type];
   }
 
   get _convertNumberToString() {
-    return convertNumberToStringByTypeMap.get(this.type);
+    return convertNumberToStringByType[this.type];
   }
 
   get _convertDateToString() {
-    return serializeDateByTypeMap.get(this.type);
+    return serializeDateByType[this.type];
   }
 
   get _convertStringToDate() {
-    return convertStringToDateByTypeMap.get(this.type);
+    return convertStringToDateByType[this.type];
   }
 
   _isStepAligned(v) {
@@ -594,9 +382,8 @@ class HTMLInputElementImpl extends HTMLElementImpl {
       return null;
     }
 
-    const stringToDateConverter = this._convertStringToDate;
     const window = this._ownerDocument._defaultView;
-    const convertedValue = stringToDateConverter(this._value);
+    const convertedValue = this._convertStringToDate(this._value);
 
     if (convertedValue instanceof Date) {
       return new window.Date(convertedValue.getTime());
@@ -614,19 +401,16 @@ class HTMLInputElementImpl extends HTMLElementImpl {
       ]);
     }
 
-    const isDateObj = v !== null && isDate(v);
-
-    if (v !== null && !isDateObj) {
+    if (v !== null && !isDate(v)) {
       throw new TypeError("Failed to set the 'valueAsDate' property on 'HTMLInputElement': The provided value is " +
         "not a Date.");
     }
 
-    if (v === null || (isDateObj && isNaN(v))) {
+    if (v === null || isNaN(v)) {
       this._value = "";
     }
 
-    const convertDateToString = this._convertDateToString;
-    this._value = convertDateToString(v);
+    this._value = this._convertDateToString(v);
   }
 
   // https://html.spec.whatwg.org/multipage/input.html#dom-input-valueasnumber
@@ -651,8 +435,7 @@ class HTMLInputElementImpl extends HTMLElementImpl {
       ]);
     }
 
-    const convertNumberToString = this._convertNumberToString;
-    this._value = convertNumberToString(v);
+    this._value = this._convertNumberToString(v);
   }
 
   // https://html.spec.whatwg.org/multipage/input.html#dom-input-stepup
@@ -955,9 +738,8 @@ class HTMLInputElementImpl extends HTMLElementImpl {
   get _minimum() {
     let min = this._defaultMinimum;
     const attr = this.getAttributeNS(null, "min");
-    const convertStringToNumber = this._convertStringToNumber;
-    if (attr !== null && convertStringToNumber !== undefined) {
-      const parsed = convertStringToNumber(attr);
+    if (attr !== null && this._convertStringToNumber !== undefined) {
+      const parsed = this._convertStringToNumber(attr);
       if (!isNaN(parsed)) {
         min = parsed;
       }
@@ -968,9 +750,8 @@ class HTMLInputElementImpl extends HTMLElementImpl {
   get _maximum() {
     let max = this._defaultMaximum;
     const attr = this.getAttributeNS(null, "max");
-    const convertStringToNumber = this._convertStringToNumber;
-    if (attr !== null && convertStringToNumber !== undefined) {
-      const parsed = convertStringToNumber(attr);
+    if (attr !== null && this._convertStringToNumber !== undefined) {
+      const parsed = this._convertStringToNumber(attr);
       if (!isNaN(parsed)) {
         max = parsed;
       }
@@ -993,9 +774,8 @@ class HTMLInputElementImpl extends HTMLElementImpl {
   }
 
   get _parsedValue() {
-    const converter = this._convertStringToNumber;
-    if (converter !== undefined) {
-      return converter(this.value);
+    if (this._convertStringToNumber !== undefined) {
+      return this._convertStringToNumber(this.value);
     }
     return this.value;
   }
@@ -1042,8 +822,7 @@ class HTMLInputElementImpl extends HTMLElementImpl {
 
   // https://html.spec.whatwg.org/multipage/input.html#concept-input-min-zero
   get _stepBase() {
-    const stringToNumber = this._convertStringToNumber;
-    const parseAttribute = attributeName => stringToNumber(this.getAttributeNS(null, attributeName));
+    const parseAttribute = attributeName => this._convertStringToNumber(this.getAttributeNS(null, attributeName));
     if (this.hasAttributeNS(null, "min")) {
       const min = parseAttribute("min");
       if (!isNaN(min)) {
